Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,5 +1,5 @@
 #include<stdio.h>
-int main()
+float main()
 {
     double a;
     scanf("%lf", &a);
